requirements.txt
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
pyngrok==7.1.6
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
urllib3==2.2.1
